Image Preprocessing code and testing
NHATS data had to be refined to ensure higher accuracies and efficiencies of machine learning models. There was significant issues with data including black bounding boxes, multiple clocks on the same page, and rotated clocks. Hence OpenCV2 was used to develop a consistent 3-part preprocessing method that was consistently applied to all images.